projects/01_fyyur/starter_code/requirements.txt
alembic==1.7.6
Babel==2.9.0
click==8.0.3
Flask==2.0.2
Flask-Migrate==3.1.0
Flask-Moment==0.11.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
phonenumbers==8.12.43
postgres==4.0
psycopg2-binary==2.9.3
psycopg2-pool==1.1
python-dateutil==2.6.0
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.31
Werkzeug==2.0.2
WTForms==3.0.1